+++ Http/ConnectionError.php
@@ -1,0 +1,13 @@
+<?php
+	/**
+	 * Exception class used by the library to distinct any thrown exception from a standard PHP Exception
+	 * @extends Exception
+	 * @package Http\Request
+	 */
+	namespace Http {
+		class ConnectionError extends \Exception {
+			public function __construct($message, $code = 0, \Exception $previous = null) {
+				parent::__construct($message, $code, $previous);
+			}
+		}
+	}