+++ Http/HttpError.php
@@ -1,9 +1,9 @@
 <?php
 	/**
-	* Exception class used by the library to distinct any thrown exception from a standard PHP Exception
-	* @extends Exception
-	* @package Http\Request
-	*/
+	 * Exception used when remote answers with a non-successful HTTP code < 400
+	 * @extends Exception
+	 * @package Http\Request
+	 */
 	namespace Http {
 		class HttpError extends \Exception {
 			public function __construct($message, $code = 0, \Exception $previous = null) {